ASP.net
This generator system for any type of category and its multiple item. It manages all detail like
category with its photo, sub item of category with photo, consumers all detail, city, state,
country lists. It is also managing detail of bill amount is pending or paid status. It is also
providing a Billing module with all data of purchased item with proper consumer detail and with Billing date with printing invoice. This system developed in 3 tier architecture. It also
provides Billing system for any company worker with admin Login.